American Sign Language Alphabets Prediction
American Sign Language is a visual language that incorporates gestures, facial expressions,head movements, body language and even the space around the speaker. Hand signs are the foundation of the language. Many signs are iconic, meaning the sign uses a visual image that resembles the concept it represents. This project helps in recognizing the hand gestures of a person and helps to communicate with the person in much effective manner.